splited more shi up and make sure u knows what it returns
